publisher: refactor page fetch
Original implementation would fetch a page's state (if any) from within
the `storePage` method. This was solely due to the fact that it was the
only call that needed to known information about a specific page.
Upcoming implementation (with respect to asset support) will require
fetching page information outside of the `storePage` call. Moving the
respective implementation out of `storePage` into its own `getPage` call
and have `storePage` use the new method.

Signed-off-by: James Knight <james.d.knight@live.com>