rest: support post with files
Upcoming support to publish assets (images and downloads) as Confluence
attachments requires (when using the REST API) using Request's POST call
with the `files` argument. Adjusting the call to support an optional
files argument (for future implementation to use).

Confluence documentation indicates that using the REST API to manage
attachments, the caller must configure the `X-Atlassian-Token` header or
the request will be blocked [1][2]. Adjusting the POST call to
automatically set this header when file content is passed in.

[1]: https://developer.atlassian.com/cloud/confluence/rest/#api-content-id-child-attachment-post
[2]: https://developer.atlassian.com/server/jira/platform/form-token-handling/

Signed-off-by: James Knight <james.d.knight@live.com>