translator: correct next-child check for indent styling
When docutils block quotes are processed for indented paragraph blocks,
child values are checked to help build consistent look-and-feel when
various block levels are in play. The traverse check to detect if a node
is a first-child will directly attempt to check the first element in the
list; however, there is no guarantee that any children will exist (thus
causing an `IndexError` exception).

Adjusting the check to ensure the next-child value actually has elements
to check against.

Signed-off-by: James Knight <james.d.knight@live.com>