assets: process doctree-resolved hook for confluence builder only
This extension will register a 'doctree-resolved' with the Sphinx
application during a setup phase. The registered function will be
invoked no matter which builder is running (e.x. the `html` builder).
Before attempting to re-check for additional image assets to resolve,
only do so when the builder instance is of a `ConfluenceBuilder` type.

This should prevent issues attempting to access an `assets` instance for
builders which do not have the assets instance. For example:

    AttributeError: 'StandaloneHTMLBuilder' object has no attribute 'assets'

Signed-off-by: James Knight <james.d.knight@live.com>