translator: interpret docname from document source
A reference-related update [1] lazily passed the docname between a
builder and a translator by tracking the active docname on a builder
instance. When a translator was created, it would pull the active
docname from the builder and use it for reference and other
capabilities. This implementation is troublesome for parallel build
events as there is not guarantee that the extension's translator will be
initialized before the builder's active docname is updated to the next
document (if any).

To prevent this risk, pull the docname from the source value in the
document instance (provided to the translator) with the help of the
environment's `path2doc` utility method

[1]: 4e58d2a516755e9f78de0352f3719184e6ed5e7f

Signed-off-by: James Knight <james.d.knight@live.com>