doc: add publish subset confluence tip
The initial introduction of the `confluence_publish_subset` feature had
provided an example snippet outlining an advanced use case of using an
environment/file configuration to tweak the option's value. Since the
configuration page is not really indented for complex examples, moving
this information into a tip document as it can be still helpful for some
users.

Signed-off-by: James Knight <james.d.knight@live.com>