config: properly check for header/footer location
When the configuration check attempts to validate the existence of the
header/footer assets (if configured), the base path for the assets from
the environment's source directory is not added. This results in the
file check looking in the incorrect area.

By adding the environment's source directory to the path-check, relative
provided headers/footers can successfully pass the configuration check.
Note that paths provided in an absolute form work as expected before and
after this change.

Signed-off-by: James Knight <james.d.knight@live.com>