test: wrap docutils namespace for sphinx invokes
For the series of unit and validations tests performed with multiple
Sphinx instances, it is important to wrap the use of the Sphinx
application around the docutils namespace. This prevents collisions
for registers directives, etc. between the various Sphinx instances.

This change corrects a series of observed test warnings generated on
execution. For example:

    WARNING: while setting up extension sphinx.domains.changeset: directive 'deprecated' is already registered, it will be overridden
    WARNING: while setting up extension sphinx.domains.changeset: directive 'versionadded' is already registered, it will be overridden
    WARNING: while setting up extension sphinx.domains.changeset: directive 'versionchanged' is already registered, it will be overridden
    WARNING: while setting up extension sphinx.domains.math: role 'eq' is already registered, it will be overridden

Signed-off-by: James Knight <james.d.knight@live.com>