test: skip code block test in sphinx 1.8+ (interim)
Due to a regression in Sphinx 1.8, `linenothreshold` feature would not
work as expected causing the code block test to fail [1]. To continue
testing with older versions of Sphinx, enable the tests for pre-v1.8
environments. When the fix for `linenothreshold` has been introduced in
a stable version of Sphinx, this test can be tweaks to support running
on the respective newer versions.

[1]: https://github.com/tonybaloney/sphinxcontrib-confluencebuilder/issues/148

Signed-off-by: James Knight <james.d.knight@live.com>