translator: handle default alignment changes in sphinx v2.1+
Starting with Sphinx 2.0.x, tables and figures were default aligned to
center when an alignment was not explicitly set [1]. This also changed
in Sphinx 2.1+ where a new alignment value "default" could be used to
determine which alignment type to manage [2]. Add appropriate alignment
handling so that under Sphinx 2.1+, values are center aligned by default
when a "default" value is provided.

A prospect change will be added in the future to allow a user to
override the desired default alignment used.

[1]: https://github.com/tk0miya/sphinx/commit/fb51336f4ac05ce198ec2bd95c54c34810f7720b
[2]: https://github.com/tk0miya/sphinx/commit/107c20a11f1753a23496a57ea8f04178e59f6810

Signed-off-by: James Knight <james.d.knight@live.com>