.travis.yml: only build on master branch
The current Travis CI configuration will trigger builds when a new
branch makes a change and when a pull request is submitted. This
results, in most cases, duplicate builds that are not necessarily
desired. Minimizing this amount of jobs on Travis CI by only building
when changes are pushed into the `master` branch and when pull request
events occur.

Signed-off-by: James Knight <james.d.knight@live.com>