Addresses #143
Adds a jira and jira_issue directive to act as a wrapper to the JIRA
structured macro available in Confluence.

Allows for configuring sets of named JIRA servers, not providing servers
at all, and a combination of named servers and overriding their
respective values on the directives.

Provides support for both single issue links (via jira_issue) and JQL
links (via jira). Allows for specifying the parameters available to
the macro.

Allows the directives to be disabled via the 'confluence_adv_restricted_macros'
configuration value

Signed-off-by: Colin Starner <colin.starner@gmail.com>