translator: force clear on figure containers
Documents using the figure directive will result in a "container" to
help visually group the image, caption and description together. In the
event where a figure has an alignment type and is very large, the image
will undesirably float outside the container area. Correcting this by
explicitly clearing floating elements when closing up a figure's
container.

Signed-off-by: James Knight <james.d.knight@live.com>