test: treat sphinx warnings as errors in unit tests
Adjusting unit tests building a Sphinx application to treat warning
events from processing as errors. This is primarily to help in future
code which wants to verify `SphinxWarning` exceptions are thrown in
select cases.

There may be a future case where this may cause issues where warnings
may be permitted for a test but still wishes to pass. In this event,
adjustments should be made to allow a unit test to explicitly tweak
whether or not this flag is set (or unset) for the given test.

Signed-off-by: James Knight <james.d.knight@live.com>