test: restore working state for test_config
Configuration-based tests have been broken since this extension attempt
to address docutils namespaces. Since application instances are created
in a context, the configuration tests switched from using `setUpClass`
to overriding `run` to limit the amount of Sphinx environments that
needed to be built; however, this change did not properly re-trigger the
default `run` method being overwritten. This commit addresses this.

[1]: 5da82a18723e00e300c36f6a71c8adf55dc9d8d0

Signed-off-by: James Knight <james.d.knight@live.com>