config.py: properly check for string type in subset configuration
The `confluence_publish_subset` option includes a configuration check to
ensure the type is of a non-string sequence type which contains string
entries; however, the implementation would verify a string type by
checking if the type is of `str`. This causes issues in Python 2.x since
Unicode strings are not of the `str` type. Adjusting the string check to
verify against a `basestring` type (with an appropriate fallback for
Python 2.x interpreters).

Signed-off-by: James Knight <james.d.knight@live.com>