test: allow tests to override warning failures
A previous commit [1] adjusted various tests to automatically result in
a failure if Sphinx generates a warning from a test's execution. This
change did not cause issues for the standard test, but has later on
revealed issues in the validation tests.

This commit adjusts the utility calls to permit a relaxed Sphinx
application instance which will not fail if a warning is triggered (i.e.
if a given test expects a warning to be generated).

[1]: 29a9643c503628f3ccbf1adb30be5eff6f15aad2

Signed-off-by: James Knight <james.d.knight@live.com>