introduce requests auth passthrough
Users may require more complex authentication handlers for their
respective Confluence instances. For these cases, allow a user to
directly pass in an authentication handler directly to the Request's
call. This prevents the need for this extension to keep expanding on
various authentication methods; leaving only the maintenance of simple
authentication in this implementation.

Signed-off-by: James Knight <james.d.knight@live.com>