publisher: generate api error if new-page response has no id
It has been observed by at least one user that a publishing event may
result in a new-page request to provide a successful response (i.e. http
code 200), but not include an identifier in the JSON-based response. To
help provide more helpful information to a user for this case (instead
of generating a `KeyError`).

Signed-off-by: James Knight <james.d.knight@live.com>