CONTRIBUTING.rst: style/formatting
Adjusting the style of the contributing documents to follow the "design"
in other documents (the design is unofficial, but formats the document
to be consistent with all the other documentation in this project; the
"design" can be re-evaluated if desired).

This change also adjusts the content of the document to help contain
lines within the eighty character boundary.

Signed-off-by: James Knight <james.d.knight@live.com>