CONTRIBUTING.rst: reflect a desired 12-month python 2.7.x support
It is a goal to help maintain support for Python 2.7.x. series for this
extension for at least a year after its end-of-life. The intent is to
help be more flexible for legacy systems which may begin to start their
transitions to a newer Python interpreter after EOL is reached.

Signed-off-by: James Knight <james.d.knight@live.com>