__init__: only inject mathbase for sphinx 1.8+
The `sphinx.ext.imgmath` extension was forcefully setup with this
extension to allow math-specific configuration values to be easily
accepted from a user's configuration. This generates a series of
warnings on Sphinx 1.7 (and older) since `sphinx.ext.mathbase` is not
already integrated into the Sphinx engine at this point. Instead, for
older revisions, do not attempt to include this extension automatically.
Users of these Sphinx variants will need to explicitly load all their
respective math extensions.

Signed-off-by: James Knight <james.d.knight@live.com>