README.rst: provide alternative pip install command
Users who are not familiar with `pip` and do not have the command
available in their terminal may be dissuaded in preparing their
environments at first. Since it can be common not to have `pip`
available on a command line, provide an installation example using the
Python interpreter to trigger a pip event.

Signed-off-by: James Knight <james.d.knight@live.com>