flag support for remote images
Confluence builder's recent support for automatic image detection
(asterisk; [1]) adjusts the builder to use the common
`post_process_image` method to manage these image types. This will
result in the image post-transform to download image assets when
building. This is not necessary for Confluence builder as Confluence
instances support remote images. Adjusting the builder's default
configuration to reflect this.

[1]: f19633aa082bf8fc7be4a9943429a6916d8ba791

Signed-off-by: James Knight <james.d.knight@live.com>