test: enable verbosity environment variable
When under a test environment, there is not a gracefully way to enable
Sphinx's verbosity flag -- primarily since we explicitly prepare Sphinx
instances ourselves in tests and do not set the verbosity flag.
Typically, a verbose Sphinx engine is not really required except for
advanced cases. To test development/testers to easily enable this option
(when using this repository's test options), allow them to enable
verbosity by setting the `SPHINX_VERBOSITY` option. For example:

    $ SPHINX_VERBOSITY=2 tox -e sandbox

Signed-off-by: James Knight <james.d.knight@live.com>