README.rst: promote api key in example pass configuration
The README's default configuration promotes the ideal that
`confluence_server_pass` will be a password value; however, for Cloud
users (which is expected to be more common), an API key would be used
instead. Adjusting the sample value to indicate that either an API key
or password value can be provided (for assuming users who may not opt
for reading the more detailed configuration documentation).

Signed-off-by: James Knight <james.d.knight@live.com>