__init__: disable math node migration
Sphinx adds a built-in `MathNodeMigrator` post-transform to help support
dealing with math-related capabilities with legacy translators. The
Confluence builder extension attempts to manage both newer and older
variants of math calls since it currently attempt to support various
older revisions of Sphinx. Since this extension supports a legacy
variant, Sphinx complains that this extension does not support the
newest math-block way of processing math-related content:

    Translator for <class 'sphinxcontrib.confluencebuilder.translator.ConfluenceTranslator'> does not support math_block node'. Please update your extension.

To prevent this warning from being generated, this extension will
explicitly remove the `MathNodeMigrator` post-transform if it is
detected.

Signed-off-by: James Knight <james.d.knight@live.com>