translator: default unknown code language to none
By default, an unspecified coding language will default to python (based
off of Sphinx's choice of default language). In the event where an
unsupported language for Confluence is provided to this extension (since
Confluence will not be able to render it), this extension would fallback
on a default language. This is unexpected since something like
'shell-session' would translate to 'python'. Instead, all unknown
language types will be translated to 'none'.

Signed-off-by: James Knight <james.d.knight@live.com>