translator: ignore nested references
Reference management assumed that nested references were not possible;
however, it has been observed in select scenarios (such as a toctree
referencing a header which also is a reference to another page) that
this is not always the case. Instead of asserting on such a scenario,
ignore a nested reference instead -- this extension will prioritize the
initial reference.

Note that this is a slight variation to what Sphinx (v2.2.0, at time of
testing) does for a documentation set applied to the HTML builder. It
has been observed that a `:local`: table of contents will generate a
reference directly to the content instead of the subsection of the
current document. It is assumed that this is not expected, thus this
extension will follow reStructuredText's wording which the local table
of contents will reference the subsection instead.

Signed-off-by: James Knight <james.d.knight@live.com>