translator: ensure figure's legend is properly aligned
When a figure is assigned an alignment type, it is expected that the
caption and legend contents also follows the same alignment type (based
off observed Sphinx stock builders). This commit adjusts the built
legend to be a div element with the respective text alignment value
assigned to a parent figure (if any is set).

Signed-off-by: James Knight <james.d.knight@live.com>