builder: properly build navigation references
The original navigation implementation would compile a list of previous
and next documents for the determined order of documents; however, the
references compiled did not properly take into account the relative URL
between the respective documents. Adjusting the caching to use the
builder's `get_relative_uri` call when forming this information.

Signed-off-by: James Knight <james.d.knight@live.com>