publisher: ability to watch state
When this extension takes advantage of creating new documents or editing
existing documents through the publishing feature, pages will be
automatically watched by the publishing user. This is not typically
ideal for documentation sets which may be published automatically (e.g.
pages purged and edited using multiple users can generate multiple
Emails in a publish event).

This commit adjusts the default watch state for pages to unwatched.
Publishers can override this by setting `confluence_watch` to a value of
`True` to always watch the state of published documents.

Signed-off-by: James Knight <james.d.knight@live.com>