builder: allow title overrides
This extension will pull the content for the first detected title
element as the title value to use for a Confluence page. This detection
may not always be the ideal choice for select documentation sets. In a
scenario where a user cannot or does not watch to explicitly adjust a
document for a title entry, they may opt to use an override option to
explicitly configure the builder to use a title value for a given
document. For example, if a user wishes to override the title value for
the `index` page, the following configuration can be used:

    confluence_title_overrides = {
        'index': 'Index Override',
    }

Signed-off-by: James Knight <james.d.knight@live.com>