state: automatic unique title names
Original title management would compile title entries and warn the
publishing user of any title conflicts, but this extension did not do
anything more than that. This can become an annoyance for large
documentation sets where titles may be conflicting but a publisher is
not that picky on the generated title values. Also, this may cause
confusion on publication when select publish events may fail when
publishing a child/nested page with a parent holding the original
ownership of the title (API error when mapping a page identifier with
itself as an ancestor).

Instead, along with warning a user of a title conflict between
documents, this extension will append an incrementing numerical value
to titles to allow a publishing event to progress without issue.

Note that for users who may publish subset of documents at a time, may
observe oddities where the title conflict resolution may generate a
different title name based on the active documents being processed.

Signed-off-by: James Knight <james.d.knight@live.com>