__init__: only register doctree resolve hook for builder
This extension introduced a hook for the `doctree-resolved` event to
help compile assets populated after a post-transformation stage. If a
user registers this extension without using the `confluence` builder,
the hook would still process; correcting.

This change also adjusts when the hook is registered from an instanceof
check to a type check. This is to avoid `SingleConfluenceBuilder` from
registering the hook which is not required in this context.

Signed-off-by: James Knight <james.d.knight@live.com>