README.rst: point to stable documentation
Adjusting the various documentation links to point to the `stable`
versions over the `latest` versions. This helps new users reading the
homepage of this extension and who typically install this extension
through a non-Git pip request, to have the most relevant documentation
for their installation.

Note that the building badge still points to the latest version.

Signed-off-by: James Knight <james.d.knight@live.com>