test: support custom builder in sandbox
Allow a developer invoking a sandbox event to target a specific builder
to compile the sandbox content against. This is primarily to assist in
builder between the default `confluence` builder and the recently added
`singleconfluence` builder. This also has the side benefit of also
supporting to build against the `html` and other builders for
comparison purposes.

Signed-off-by: James Knight <james.d.knight@live.com>