assets: provided fallback for asset paths
Asset-based nodes processed by this extension will either provide a URI
to a relative or absolute path. Part of the asset manager's process will
attempt to determine the absolute path of a resource for future work. In
the case for a relative path, a relative path is assumed to be located
alongside the source content; however, this is not always the case for
extensions which generate assets into the build output directory but
only provide a relative path to the asset. To help deal with this case,
when a relative path based off a documentation's source directory
cannot be found, the process will fallback to a path based off the
user-defined output directory.

Signed-off-by: James Knight <james.d.knight@live.com>