correct hook with multiple builders
For the standard builder in this extension, the `doctree-resolved`
event is observed to help populate images which may be generated when a
doctree has been resolved. When the single builder was introduced in
this extension, the builder attribute (`app.builder`) no longer appears
to be always initialized. Before type-checking the builder being used,
hook on the `builder-inited` event to ensure the builder has been
initialized.

Signed-off-by: James Knight <james.d.knight@live.com>