tox: manage supported sphinx versions
Adjusting the tox configuration to test for the more recent releases of
Sphinx. Since this extension aims to support a least five latest
releases of Sphinx, dropping tests for some older versions. Note, that
Sphinx v1.8.x series will still be tested against since it is the only
version still supported on Python 2, which will continue to have support
for at least a couple of more months.

Signed-off-by: James Knight <james.d.knight@live.com>