directives: refactor metadata directive
The initial implementation of ConfluenceMetadataDirective most likely
used the existing JIRA directives as a reference for implementation. The
JIRA directive implementation uses a base directive to help deal with
various types of JIRA directive definition. The metadata directive also
defined its own base directive; however, (at this time) there does not
appear to be a need for a base directive as there is only a single use
for this type. Refactoring the implementation now to be a bit smaller.

This change also sorts the directive to the top of the page.

Signed-off-by: James Knight <james.d.knight@live.com>