rework label feature restrictor
The initial implementation for page label support uses the
macro-restrictor feature to disable its use; however, this feature is
not a macro. At the same time, label capabilities could be disabled by
an administrator as well. Therefore, it would still be ideal if a
publisher had the ability to prevent an attempt to adjust labels on a
page.

The first part of this change tweaks the advanced configuration item
`confluence_adv_restricted_macros` which will be renamed to
`confluence_adv_restricted` to be more generic. Since this is an
advanced/undocumented feature, there is less concern for configuration
changes between revisions of this extension (however, it may be ideal to
note the change in the changelog when a new release is made).

With the restriction feature being more generic, we can use the simple
tag/quirk list to help disable the label interaction. The original tag
created was named `confluence_metadata`; however, this will be changed
to `labels` to be more specific on what is being restricted.

The final change will deal with where the restrictions take place. The
original implementation would only disable accepting the new label
content at the page-level. This would still result in the publish
attempt to query for labels and attempt to re-set labels (which is not
really required and may cause an API failure if labels are restricted).
Instead, the translator will also accept new label information when
processing a document, but the restriction flag will be used to prevent
a publishing event from query and attempting to set label information.

Signed-off-by: James Knight <james.d.knight@live.com>