support per-document labels
The following adjusts the label mechanics to have the builder
implementation process all prospect metadata information for a document
instead of passing metadata information between a translator and builder
via configuration options.

This change is being made to first prevent the translator from
manipulating the configuration context (which is not necessarily ideal).
The original request for label support outlined a desired to have a
per-page label assignment. This commit brings this capability by now
tracking metadata information at a per-document level (instead of a
single last-read metadata to be applied to all pages).

Note that this change does prevent the ability for a user to apply a
"global" label set to all pages, which was implicitly capable before. A
future commit can be made to allow setting global labels via a
configuration option.

Note this change also introduces some corrections/improvements to the
metadata node processing. First, this change provides a more flexible
parsing of label data provide from a document -- this includes
supporting multilines of labels and dealing with additional whitespaces.
Second, a correction has been made to ensure node information can be
properly shared when assembling a single document for the
`singleconfluence` builder. The directive now populates the attribute
container which allows stored information to be copied between a cloned
node. The implementation issue exists in the existing JIRA-based
implementation, which this implementation appears to be based off of
(corrections to the JIRA implementation should be made in the future).

Signed-off-by: James Knight <james.d.knight@live.com>