storage: do not inject newline inside attachment tag
It has been observed that (in at least) Confluence Cloud that when a
newline is injected in the `ri:attachment` tag, typically only when
`ri:content-title` is used, that Confluence will report a 500 error when
publishing. This commit removes the suffix injection (from an explicit
newline to an implicit empty/none) and publish events appear to work as
expected.

Signed-off-by: James Knight <james.d.knight@live.com>