builder: handle sphinx.ext.inheritance_diagram nodes
Provide support for an internal Sphinx extension `inheritance_diagram`
by pre-processing diagrams and replaced with respective images.
Typically, node support from `sphinx.ext.inheritance_diagram` would be
added to the builder; however, this extension renders graphs during the
translation phase (which is not ideal for how assets are managed in this
extension).

Instead, this implementation just traverses for inheritance diagrams,
generates renderings and replaces the nodes with image nodes (which in
turn will be handled by the existing image-based implementation).

Signed-off-by: James Knight <james.d.knight@live.com>