tox: correct flake8 invoke
The lint environment invoking `flake8` was broken ages ago [1] when the
argument was interpreted as a namespace value instead of a path
argument. Adjusting the environment to properly test this project's
implementation.

Additional tweaks include renaming the `lint` environment to `flake8`
(to be explicit in what is being used) as well as also including the
testing implementation to be verified as well.

A series of personally preferred ignore options have been added into the
ignore list. In the event that an option is desired to be added/removed
from this list, it can be discussed in an issue/pull request.

[1]: 71265474065083f5b4b6201c6fe8278a79bfa8cb

Signed-off-by: James Knight <james.d.knight@live.com>