set explicit except catches
With flake8 functionality restored [1], addressing linting errors in the
project. This commit sets explicit exceptions to catch instead of the
non-recommended catch-all method.

An exception has been made in the rest's format-error call, just to
ensure all bad data use-case does not cause error which could cause a
user's failure case to hard-fail. This can be looked at a later time to
find the explicit exceptions which should be caught here.

[1]: 1a73ce8223c25a5816e1a153c7dfbba17a96bfdc

Signed-off-by: James Knight <james.d.knight@live.com>