test: validate admonitions with parser
Providing an initial test case which switches from using diff-comparing
output and expected files validation to instead using a parser. This is
using the Beautiful Soup library to assist in examining generating
output instead of having to rely on manage sets of "excepted" files
which sometimes undesirably need to be updated for trivial reasons.

Signed-off-by: James Knight <james.d.knight@live.com>