tox: manage supported sphinx versions (sphinx 3.2.x)
Adjusting the tox configuration to test for the more recent releases of
Sphinx. Since this extension aims to support a least five latest
releases of Sphinx, dropping tests for some older versions. Note, that
Sphinx v1.8.x series will still be tested against since it is the only
version still supported on Python 2 (which is still currently
supported).

Signed-off-by: James Knight <james.d.knight@live.com>