README.rst: default link to stable documentation
The README document provides a build status badge which can be used to
quickly jump to this extension's documentation. While it is desired to
reflect the most recent documentation's build status, users may assume
that the documentation linked is the most recent version available via
PyPI sources (which is not the case). To help prevent confusion for
users, link to the stable documentation by default.

Signed-off-by: James Knight <james.d.knight@live.com>