storage: correct literal parsing newlines
When processing each text node's contents, newlines are stripped to
ensure horizontal view is taken advantage of. This does not apply to
parsed literals which expect a matching format. The split of translator
implementation introduces an issue where it would set an internal
"is-parser-literal" flag when the processing nodes were not actually a
parsed literal type; correcting.

Signed-off-by: James Knight <james.d.knight@live.com>