support leading title link support from toc directive
When using restructuredText's table of contents directive (`contents`),
the reference to the leading section page was always removed when the
first section (the "title") was removed (by default) from a document.
This was to prevent generating an entry point in the TOC which would not
have a proper link target to reference. This causes two issues:

1) This extension would modify a possible expected output to have an
    anchor link to the top of a document.
2) If a table of contents had multiple child entries to display, the
    output would generate a nested list category for a "removed" root
    which would produce an empty/undesired bullet point.

Instead, this commit introduces support for the reference missing the
target by detecting that a reference is now a "#top" reference in the
builder. This in turn allows during the translation phase to attempt to
build a proper link. In the case for the storage format translator, an
anchor link to "#top" can be created.

Signed-off-by: James Knight <james.d.knight@live.com>