improve support for section numbering
The following provides a more complete for section numbering support
when used with this extension. This extension's builders never built
proper targets for sections to reference against. Section names
identifiers are internally managed by the Confluence instance and the
target mappings never took into consideration of a section with a
prefixed section number. Cached target entries will now handle when a
section number is added to a section text.

In addition, the singleconfluence builder required implementing its own
title cache list, as the section targets need to be compiled after
assembling the section numbers, but before the primary doctree is built.
This is a result of pre-merged doctree's which will point to
page-specific targets (e.g. `sub/#target`) but the original
implementation would assume only an index-based target existed (e.g.
`index/#target`) -- which could result in conflicts.

Signed-off-by: James Knight <james.d.knight@live.com>