test: adding additional section numbering checks
The following increases a number of sanity checks to make when verifying
proper reference building when section numbering is involved. This
includes ensuring proper anchor points are created for documents using
a `toctree` directive's `:numbered:` option, along with a `:contents:`
directive for page-contained TOCs.

Testing also includes verifying the use of an injecting `#top` link for
a content's leading link, when a document's title entry is removed (by
default) during processing.

Signed-off-by: James Knight <james.d.knight@live.com>