storage: handle anchor-specific targets flagged as internal
In select cases, it has been observed that a reference node may contain
a `refuri` definition which contains an anchor to the active document
being processed. The existing implementation would only be able to
handle this case if the reference was also not marked as internal;
however, it has been observed that in select cases (such as using the
`sphinx.ext.todo` extension) that local anchor entries can also be on
references which have an internal flag set.

This commit re-works the process `refuri` process flow by:

1) Always checking for an local anchor reference to forward to to the
    internal identifier-based reference implementation.
    `_visit_reference_intern_id` (the fix).
2) If not internal, continue to forward it to the external reference
    implementation.
3) Otherwise, process it as an internal URI.

Signed-off-by: James Knight <james.d.knight@live.com>