inject autosummary nodes after a proper extension check
This extension would attempt to register compatible autosummary nodes
when the `sphinx.ext.autosummary` extension was loaded. This would allow
a user to take advantage of the autosummary capabilities with this
extension.

The problem with the check on `app.config.extensions` is that that
configuration file only indicates extensions the user wishes to
explicitly load, and not what extensions have been actually loaded into
the Sphinx engine.

The injection process has been changed to better handle autosummary
support:

1) Explicitly check against the `app.extensions` loaded extensions to
    determine if the nodes should be injected.
2) Only attempt to inject after the `builder-inited` event to ensure
    that modules loaded after the Confluence builder extension which may
    load `sphinx.ext.autosummary` extension themselves does not cause an
    missed injection.

Signed-off-by: James Knight <james.d.knight@live.com>