.travis.yml: drop travis-ci
Changes at Travis CI now require making a support request for OSS
credits to take advantage of their services. While it is possible to
attempt to make a format request, it is unknown on the turnaround time
and attempting to pre-calculate expected credit usage to request in the
inquiry.

It appears easier to switch to GitHub actions (since this project's
respository lives on GitHub) and get the same results needed for this
extension.

Signed-off-by: James Knight <james.d.knight@live.com>