storage: default doctitle for images for non-publishing documents
In the event where a document cannot be published due to a lack of a
title value, this can result in image nodes causing an exception since
there is no title value to process:

     File ".\sphinxcontrib\confluencebuilder\translator\storage.py", line 1262, in visit_image
        hosting_doctitle = self._escape_sf(hosting_doctitle)
      File ".\sphinxcontrib\confluencebuilder\translator\storage.py", line 2019, in _escape_sf
        data = data.replace('&', '&amp;')
    AttributeError: 'NoneType' object has no attribute 'replace'

For the output of documents which do not have a title entry to work
with, there is still a desire to (at least) continue processing the
documentation set. Documents in this scenario will have the storage
format content with the desired attachment entries (for inspection and
other possible user actions).

Signed-off-by: James Knight <james.d.knight@live.com>